+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @opencx/widget-core
 
+## 4.0.8
+
+### Patch Changes
+
+- add WidgetConfig.chatBannerItems
+
 ## 4.0.7
 
 ### Patch Changes