+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @opencx/widget-react
 
+## 4.0.8
+
+### Patch Changes
+
+- add WidgetConfig.chatBannerItems
+- Updated dependencies
+  - @opencx/widget-core@4.0.8
+  - @opencx/widget-react-headless@4.0.8
+
 ## 4.0.7
 
 ### Patch Changes